tests - export params, alter table refactor
