+++ phpunit.xml.dist
@@ -6,8 +6,8 @@
          convertNoticesToExceptions="true"
          convertWarningsToExceptions="true"
          processIsolation="false"
-         stopOnFailure="true"
-         stopOnError="true"
+         stopOnFailure="false"
+         stopOnError="false"
          syntaxCheck="false"
          bootstrap="tests/bootstrap.php">
